+++ Plugins/Events/Events/InputEvents.cpp
@@ -27,6 +27,10 @@ InputEvents::InputEvents(Services::HooksProxy* hooker)
     Events::InitOnFirstSubscribe("NWNX_ON_INPUT_FORCE_MOVE_TO_OBJECT_.*", [hooker]() {
         hooker->RequestSharedHook<API::Functions::_ZN12CNWSCreature27AddMoveToPointActionToFrontEt6Vectorjjiffiiiiii, int32_t>(&AddMoveToPointActionToFrontHook);
     });
+
+    Events::InitOnFirstSubscribe("NWNX_ON_INPUT_CAST_SPELL_.*", [hooker]() {
+        hooker->RequestSharedHook<API::Functions::_ZN12CNWSCreature19AddCastSpellActionsEjiiii6Vectorjiiihiiih, int32_t>(&AddCastSpellActionsHook);
+    });
 }
 
 void InputEvents::HandlePlayerToServerInputWalkToWaypointHook(bool before, CNWSMessage *pMessage, CNWSPlayer *pPlayer)
@@ -80,6 +84,25 @@ void InputEvents::AddMoveToPointActionToFrontHook(bool before, CNWSCreature *pCr
 
         Events::SignalEvent(before ? "NWNX_ON_INPUT_FORCE_MOVE_TO_OBJECT_BEFORE" : "NWNX_ON_INPUT_FORCE_MOVE_TO_OBJECT_AFTER", pCreature->m_idSelf);
     }
+}
+
+void InputEvents::AddCastSpellActionsHook(bool before, CNWSCreature *pCreature, uint32_t nSpellId, int32_t nMultiClass, int32_t nDomainLevel,
+        int32_t nMetaType, int32_t bSpontaneousCast, Vector vTargetLocation, Types::ObjectID oidTarget, int32_t bAreaTarget, int32_t bAddToFront,
+        int32_t bFake, uint8_t nProjectilePathType, int32_t bInstant, int32_t bAllowPolymorphedCast, int32_t nFeat, uint8_t nCasterLevel)
+{
+    (void)nDomainLevel; (void)nMetaType; (void)bSpontaneousCast; (void)bAddToFront; (void)bFake;
+    (void)nProjectilePathType; (void)bInstant; (void)bAllowPolymorphedCast; (void)nFeat; (void)nCasterLevel;
+
+    Events::PushEventData("TARGET", Utils::ObjectIDToString(oidTarget));
+    Events::PushEventData("SPELL_ID", std::to_string(nSpellId));
+    Events::PushEventData("MULTICLASS", std::to_string(nMultiClass));
+
+    Events::PushEventData("IS_AREA_TARGET", std::to_string(bAreaTarget));
+    Events::PushEventData("POS_X", std::to_string(vTargetLocation.x));
+    Events::PushEventData("POS_Y", std::to_string(vTargetLocation.y));
+    Events::PushEventData("POS_Z", std::to_string(vTargetLocation.z));
+
+    Events::SignalEvent(before ? "NWNX_ON_INPUT_CAST_SPELL_BEFORE" : "NWNX_ON_INPUT_CAST_SPELL_AFTER", pCreature->m_idSelf);
 }
 
 }